Initial commit
In this first commit, I've moved over the interesting code from the
diffux repo, created a gemspec file, introduced rspec and verified that
specs are running.

In its essence this should be seen as a pure move. I did change a couple
of lines, but that was only to make sure all the dependencies were in
place. There were two references to `Rails` that I had forgotten to
remove in previous cleanup commits in the diffux repo. But I figured
those were small enough to just bake in to this commit.