Add `diffux-compare` command-line tool
This command-line tool takes two images and diffs them. If a diff is
found, a diff.png file is saved in the current folder (you can override
this). Plus the diff is outputted as "DIFF: 22.0%". If no diff is found,
no image is saved and nothing is logged (exit status 0).