Add ability to limit the snapshot by a query selector
This commit adds the `crop_selector` feature to Diffux::Snapshotter. By
specifying such an option when taking a snapshot, only that box will be
saved in the snapshot image. Here's an example (using the command-line
tool) grabbing only the "Why does it work?" section from causes.com:

    diffux-snapshot --url https://www.causes.com \
                    --outfile causes.png
                    --cropselector '.why'

This is useful if you for instance have a main column with all the good
stuff and the sidebar is full of ads and other things you can't control.
Or you might have a component that you are working on and want to take a
snapshot before and after a change you're making.