Add Code Climate test coverage reporting
In the diffux project we have been using coveralls for test coverage
reporting. However, we have noticed in Travis CI that for Ruby 2.1.0,
our tests are "passing" even if they have failures. While we haven't
taken the time to figure out exactly what is going on there yet, we have
a hunch that it could be coveralls exiting with a success code in this
case which is masking the failed specs.

Additionally, since we are already using Code Climate for code quality
and security analysis, we think that it may be convenient to have code
coverage in the same place. This commit sets that up so Travis CI will
automatically push code coverage to our Code Climate repo.